% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_snp_clusters.R
\name{assign_snp_clusters}
\alias{assign_snp_clusters}
\title{Assign SNP Clusters to Samples}
\usage{
assign_snp_clusters(
  seqs = NULL,
  collectdts,
  snp_matrix = NULL,
  threshold = 5,
  clockrate = -1,
  transm_rate = -1
)
}
\arguments{
\item{seqs}{Matrices containing sequences.}

\item{collectdts}{A numeric vector of collection dates, with names corresponding
to the rownames of the \code{seqs} matrix.}

\item{snp_matrix}{If SNP distances have been previously computed, include
this argument to avoid re-computing.}

\item{threshold}{A numeric value of the SNP threshold, or the transmission
threshold if using transmission clusters (see \code{transcluster} for more info).}

\item{clockrate}{A numeric value of the clock rate if using transmission
clusters.}

\item{transm_rate}{A numeric value of the transmission rate if using
transmission clusters.}
}
\value{
A data frame containing sample ID, cluster name, and collection date
for the sample.
}
\description{
This function assigns SNP clusters to samples based on the SNP distance and
some threshold, or, alternatively, based on transmission clusters as in the
Stimson et al. paper, \href{https://academic.oup.com/mbe/article/36/3/587/5300248}{Beyond the SNP Threshold: Identifying Outbreak Clusters Using Inferred Transmissions}.
}
