% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_BEAST2_clusters.R
\name{create_BEAST2_clusters}
\alias{create_BEAST2_clusters}
\title{Create Clusters for BEAST2 Input}
\usage{
create_BEAST2_clusters(
  seqs,
  cluster_assignments,
  min_cluster_size = 4,
  min_varsites = 8,
  snps_only = TRUE,
  constant_sites = "",
  cluster_dictionary_file = NULL,
  fasta_dir = NULL
)
}
\arguments{
\item{seqs}{A matrix of DNA sequences, where rownames correspond
to sample IDs, or an \code{ape::DNAbin} object.}

\item{cluster_assignments}{A data frame of cluster assignments, including
at least \code{sample_id} and \code{cluster_name}, as output by \code{assign_snp_clusters}.
Must also include \code{collectdt} if writing a cluster dictionary.}

\item{min_cluster_size}{The number of samples required to keep a cluster.
If the cluster is too small, discard it.}

\item{min_varsites}{The number of variable sites required to keep a cluster.
If there aren't enough variable sites, discard the cluster.}

\item{snps_only}{If TRUE, keep only SNPs in the output DNAbin objects.}

\item{constant_sites}{A string of constant sites to add to the beginning of
each sequence. This is necessary for BEAST2 input if using only SNPs. Defaults
to an empty string.}

\item{cluster_dictionary_file}{Optionally, a file path to write a cluster
dictionary to, containing cluster names, sizes, and varsites.}

\item{fasta_dir}{Optionally, a directory to write FASTA files to for each
cluster. FASTA files are automatically named after the cluster names.}
}
\value{
A named list of matrices containing SNPs.
}
\description{
Using SNP cluster information and DNA sequences, separate sequences into
separate DNAbin objcts by cluster. If desired, keep only SNPs.
}
