% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess_checks.R
\name{ess_checks}
\alias{ess_checks}
\title{Use Effective Sample Size to Check Mixing}
\usage{
ess_checks(
  program = c("BEAST2", "TransPhylo"),
  path_to_mcmc_log,
  burn_in_fraction = 0.1,
  sample_interval = 5000,
  min_ess
)
}
\arguments{
\item{program}{Character string. Either "BEAST2" or "TransPhylo".}

\item{path_to_mcmc_log}{Character string. Path to the MCMC log file.}

\item{burn_in_fraction}{Numeric. Fraction of the MCMC chain to discard as
burn-in.}

\item{sample_interval}{Numeric. Interval at which samples were taken.}

\item{min_ess}{Numeric. Minimum effective sample size to consider acceptable.}
}
\description{
For a BEAST2 run or a TransPhylo run, pull the effective sample size
for estimated parameters. Check that they are above some threshold.
}
