% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cli_beast2_programs.R
\name{get_mcctree}
\alias{get_mcctree}
\title{Get Maximum Clade Credibility Trees Using Command-Line TreeAnnotator}
\usage{
get_mcctree(
  input_treesfile,
  output_dir,
  beast_iterations = 2e+07,
  burnin_fraction = 1/2,
  heights = "CA",
  treeannotator_path = "/Applications/\\"BEAST 2.7.7\\"/bin/treeannotator"
)
}
\arguments{
\item{input_treesfile}{Character string. Path to the BEAST2 .trees file.}

\item{output_dir}{Character string. Path to the desired output directory
for the MCC tree nexus file.}

\item{beast_iterations}{Numeric. Number of iterations in the BEAST2 run.}

\item{burnin_fraction}{Numeric. Fraction of the MCMC chain to discard as
burn-in.}

\item{heights}{Character string. Node heights for the tree. Beware that if
using anything other than common ancestor, could get some weird trees that
won't work with TransPhylo.}

\item{treeannotator_path}{Character string. Path to the TreeAnnotator
executable.}
}
\description{
Run TreeAnnotator to get the maximum clade credibility tree from a BEAST2
tree file.
}
