% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TransPhylo.R
\name{get_prob_source}
\alias{get_prob_source}
\title{Get Infector Probabilities from TransPhylo Results}
\usage{
get_prob_source(tp_res, type = c("mcctrees", "trees_sample"))
}
\arguments{
\item{tp_res}{A list of resTransPhylo objects.}

\item{type}{A character string specifying the type of trees that were
analyzed. Can be "mcctrees" or "trees_sample".}
}
\value{
A data frame containing the probability of each sample being an
infection source.
}
\description{
Get Infector Probabilities from TransPhylo Results
}
