% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{Run Linear or Logistic Regression with TransPhylo Results}
\usage{
regression(
  method = c("logistic", "linear"),
  cleaned_data,
  prob_source,
  prob_cutoff = NULL
)
}
\arguments{
\item{method}{A character string specifying the type of regression to run.
Can be "linear" or "logistic".}

\item{cleaned_data}{A data frame containing the covariates to use in the
regression. Must contain a column "SampleID" with the sample IDs.}

\item{prob_source}{A vector of probabilities output by TransPhylo, indicating
the probability of each sample being the infection source.}

\item{prob_cutoff}{A numeric specifying the probability cutoff to use for
logistic regression. If NULL, logistic regression will not be run.}
}
\value{
An \code{lm} or \code{glm} object.
}
\description{
This function takes probability estimates from TransPhylo and runs a
linear regression on them, or logistic regression if a probability cutoff
to label an individual as an infection source is specified.
}
