% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cli_beast2_programs.R
\name{run_beast2}
\alias{run_beast2}
\title{Run Command-Line BEAST2}
\usage{
run_beast2(
  input_xml_path,
  beast2_path = "/Applications/\\"BEAST 2.7.7\\"/bin/beast"
)
}
\arguments{
\item{input_xml_path}{Character string. Path to the BEAST2 XML file.}

\item{beast2_path}{Character string. Path to the BEAST2 executable.}
}
\description{
Run BEAST2 from an XML file. Tree files, log files, etc. will be written
to the same directory as the XML file.
}
