% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_BEAST2_trees.R
\name{sample_BEAST2_trees}
\alias{sample_BEAST2_trees}
\title{Sample BEAST2 Trees}
\usage{
sample_BEAST2_trees(trees_file, n_trees, seed = NULL, out_dir = NULL)
}
\arguments{
\item{trees_file}{A string of the path to the BEAST2 posterior trees file.}

\item{n_trees}{An integer of the number of trees to sample.}

\item{seed}{Optionally, a seed to use for tree sampling.}

\item{out_dir}{Optionally, a string of the directory to write the sampled trees to.}
}
\value{
An \code{ape} multiPhylo object containing the tree sample.
}
\description{
This function takes BEAST2 posterior trees and samples a subset of them,
making sure to keep tree names.
}
