% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{Run Linear or Logistic Regression with TransPhylo Results}
\usage{
regression(
  method = c("logistic", "linear", "bayesian_logistic_misclass"),
  cleaned_data,
  prob_source,
  prob_cutoff = NULL,
  sensitivity = NULL,
  specificity = NULL,
  ...
)
}
\arguments{
\item{method}{A character string specifying the type of regression to run.
Can be "linear", "logistic", or "bayesian_logistic_misclass".}

\item{cleaned_data}{A data frame containing the covariates to use in the
regression. Must contain a column "SampleID" with the sample IDs.}

\item{prob_source}{A data frame of probabilities output by TransPhylo with columns
\code{SampleID} and \code{prob_source} (as output by \code{run_TransPhylo}), indicating
the probability of each sample being the infection source.}

\item{prob_cutoff}{A numeric specifying the probability cutoff to use for
logistic regression. If NULL, logistic regression will not be run.}

\item{sensitivity}{Numeric. Only used if \code{method} is "bayesian_logistic_misclass".}

\item{specificity}{Numeric. Only used if \code{method} is "bayesian_logistic_misclass".}

\item{...}{Additional arguments to pass to Stan if \code{method} is
"bayesian_logistic_misclass". Iterations and chains default to 2000 and 4
respectively.}
}
\value{
An \code{lm} object, a \code{glm} object, or a \code{stanfit}.
}
\description{
This function takes probability estimates from TransPhylo and runs a
linear regression on them, or logistic regression if a probability cutoff
to label an individual as an infection source is specified.
}
