% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_TransPhylo.R
\name{run_TransPhylo}
\alias{run_TransPhylo}
\title{Run TransPhylo on a Set of Trees}
\usage{
run_TransPhylo(
  trees,
  cluster_name = NULL,
  type = c("mcctrees", "trees_sample"),
  cluster_dict,
  out_dir = "TransPhylo",
  output_name = "tp_res",
  ...
)
}
\arguments{
\item{trees}{An \code{ape} multiPhylo object containing the trees to be
analyzed. If \code{type} is "mcctrees", this should be all MCC trees for
all clusters, and the names of the trees need to match the names
of the trees in the cluster dictionary. If \code{type} is "trees_sample",
this should be a sample of posterior BEAST2 trees for a single cluster,
and the cluster name should be provided as an argument.}

\item{cluster_name}{If type is "trees_sample", this should be the name
of the cluster for which the trees are being analyzed. This should match
the cluster name in the cluster dictionary. If type is "mcctrees", this
argument is ignored.}

\item{type}{A character string specifying the type of trees to be
analyzed. Can be "mcctrees" or "trees_sample". If "mcctrees", this
function will run TransPhylo on all MCC trees for all clusters. If
"trees_sample", this function will run TransPhylo on a sample of
posterior BEAST2 trees for a single cluster.}

\item{cluster_dict}{A data frame containing \code{sample_id}, \code{cluster_name},
and \code{collectdt} for each sequence, as output by \code{assign_snp_clusters}.}

\item{out_dir}{A character string specifying the directory where
the TransPhylo results will be saved.}

\item{output_name}{A character string specifying the name of the output
file.}

\item{...}{Additional arguments to be passed to
\code{TransPhylo::infer_multittree_share_param}.}
}
\value{
A data frame containing the probability of each sample being an
infection source.
}
\description{
This is a wrapper for TransPhylo::infer_multittree_share_param.
Can use it to run TransPhylo on a set of trees, either all MCC trees
with parameter sharing, or a sample of posterior BEAST2 trees without
parameter sharing (i.e., a sample of BEAST2 posterior trees for a single
cluster, to incorporate some phylogenetic uncertainty).
}
